Unified download and upload scripts
