+++ packages/ckeditor5-dev-env/tests/translations/utils.js
@@ -5,14 +5,45 @@
 
 'use strict';
 
+const sinon = require( 'sinon' );
 const chai = require( 'chai' );
 const expect = chai.expect;
+const mockery = require( 'mockery' );
 
 describe( 'dev-env/translations/utils', () => {
-	let utils;
+	let stubs, utils;
 
 	beforeEach( () => {
+		mockery.enable( {
+			useCleanCache: true,
+			warnOnReplace: false,
+			warnOnUnregistered: false
+		} );
+
+		stubs = {
+			chalk: {
+				cyan: sinon.stub().callsFake( msg => msg )
+			},
+			logger: sinon.stub().returns( {
+				info: sinon.stub(),
+				warning: sinon.stub(),
+				error: sinon.stub(),
+				_log: sinon.stub()
+			} )
+		};
+
+		mockery.registerMock( 'chalk', stubs.chalk );
+		mockery.registerMock( '@ckeditor/ckeditor5-dev-utils', {
+			logger: stubs.logger
+		} );
+
 		utils = require( '../../lib/translations/utils' );
+	} );
+
+	afterEach( () => {
+		sinon.restore();
+		mockery.deregisterAll();
+		mockery.disable();
 	} );
 
 	describe( 'verifyProperties()', () => {
@@ -100,6 +131,44 @@ describe( 'dev-env/translations/utils', () => {
 					foo: () => {}
 				}, [ 'foo' ] );
 			} ).to.not.throw( Error );
+		} );
+	} );
+
+	describe( 'createLogger()', () => {
+		it( 'should be a function', () => {
+			expect( utils.createLogger ).to.be.a( 'function' );
+		} );
+
+		it( 'should return an object with methods', () => {
+			const logger = utils.createLogger();
+
+			expect( logger ).to.be.an( 'object' );
+			expect( logger.progress ).to.be.a( 'function' );
+			expect( logger.info ).to.be.a( 'function' );
+			expect( logger.warning ).to.be.a( 'function' );
+			expect( logger.error ).to.be.a( 'function' );
+			expect( logger._log ).to.be.a( 'function' );
+		} );
+
+		it( 'should call the info method for a non-empty progress message', () => {
+			const logger = utils.createLogger();
+
+			logger.progress( 'Example step.' );
+
+			expect( logger.info.callCount ).to.equal( 1 );
+			expect( logger.info.firstCall.args[ 0 ] ).to.equal( '\n📍 Example step.' );
+			expect( stubs.chalk.cyan.callCount ).to.equal( 1 );
+			expect( stubs.chalk.cyan.firstCall.args[ 0 ] ).to.equal( 'Example step.' );
+		} );
+
+		it( 'should call the info method with an empty message for an empty progress message', () => {
+			const logger = utils.createLogger();
+
+			logger.progress();
+
+			expect( logger.info.callCount ).to.equal( 1 );
+			expect( logger.info.firstCall.args[ 0 ] ).to.equal( '' );
+			expect( stubs.chalk.cyan.called ).to.equal( false );
 		} );
 	} );
 } );
